backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
sqlmodel==0.0.14
google-generativeai==0.3.2
pypdf==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
